:construction: Public API cleanup
The way that I initially wrote the methods mean that having the `generateSeed()` method as a separate call means that the seed is static, and even if I increment the counter in the `generatePassword()` method, it still uses the same static seed, regardless of the passed counter value.

So `generateSeed()` is now a helper function.